pyproject.toml
[project]
name = "rag-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "google-cloud-aiplatform>=1.138.0",
    "google-genai>=1.64.0",
    "google-generativeai>=0.8.6",
    "pinecone>=8.0.1",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
